bug fixes. made all js files modules (all artists and all albums still not working)

+++ src/main/js/fullAlbum.js
@@ -1,17 +1,11 @@
 // get the album id from the url
-import {getCsrfInfo} from "../modules/csrf";
+import {getCsrfInfo} from "./modules/csrf";
 
-let url = window.location.href;
-let albumId = url.substring(url.lastIndexOf("/") + 1);
-
-if (albumId.indexOf("?") !== -1) {
-    albumId = albumId.substring(0, albumId.indexOf("?"));
-}
 
 const {header, token} = getCsrfInfo();
 
-export function deleteAlbum(albumId) {
-    fetch(`/api/album/${albumId}`, {
+export function deleteAlbum(albId) {
+    fetch(`/api/album/${albId}`, {
         method: "DELETE",
         headers: {
             Accept: "application/json",
@@ -27,37 +21,45 @@ export function deleteAlbum(albumId) {
     });
 }
 
-export function getAlbum(albumId) {
-    return fetch(`/api/album/${albumId}`)
+export function getAlbum(albId) {
+    return fetch(`/api/album/${albId}`)
         .then((response) => response.json())
         .then((album) => {
             return album;
         });
 }
 
-export function getAlbumsArtists(albumId) {
-    return fetch(`/api/artist/album/${albumId}/artists`)
+export function getAlbumsArtists(alId) {
+    return fetch(`/api/artist/album/${alId}/artists`)
         .then((response) => response.json())
         .then((artists) => {
             return artists;
         });
 }
 
-export function getAlbumsSongs(albumId) {
-    return fetch(`/api/song/album/${albumId}/songs`)
+export function getAlbumsSongs(alId) {
+    return fetch(`/api/song/album/${alId}/songs`)
         .then((response) => response.json())
         .then((songs) => {
             return songs;
         });
 }
 
-let deleteAlbumButton = document.getElementsByClassName("delete-button")[0];
+if (window.location.href.indexOf("fullAlbum") !== -1) {
+    let deleteAlbumButton = document.getElementsByClassName("delete-button")[0];
+    let url = window.location.href;
+    let albumId = url.substring(url.lastIndexOf("/") + 1);
+    if (albumId.indexOf("?") !== -1) {
+        albumId = albumId.substring(0, albumId.indexOf("?"));
+    }
+    let songsList = document.getElementById("album-songsList");
 
-let album = getAlbum(albumId);
-let albumsArtists = getAlbumsArtists(albumId);
-let albumsSongs = getAlbumsSongs(albumId);
+    let album = await getAlbum(albumId);
+    let albumsArtists = await getAlbumsArtists(albumId);
+    let albumsSongs = await getAlbumsSongs(albumId);
+    let artistsList = document.getElementById("album-artistsList");
 
-album.then((album) => {
+
     console.log(album);
     document.getElementById("album-username").innerHTML = album.username;
     document.getElementById("album-albumName").innerHTML = album.albumName;
@@ -66,11 +68,8 @@ album.then((album) => {
         album.officialTrackCount;
     document.getElementById("album-albumStatus").innerHTML = album.albumStatus;
     document.getElementById("album-releaseDate").innerHTML = album.releaseDate;
-});
 
-albumsArtists.then((artists) => {
-    let artistsList = document.getElementById("album-artistsList");
-    artists.forEach((artist) => {
+    for (let artist of albumsArtists) {
         let artistRow = document.createElement("tr");
         let artistName = document.createElement("td");
         artistName.innerText = artist.name;
@@ -85,13 +84,10 @@ albumsArtists.then((artists) => {
             window.location.href = artistRow.getAttribute("data-href");
         });
         artistsList.appendChild(artistRow);
-    });
-});
+    }
 
-albumsSongs.then((songs) => {
-    console.log(songs);
-    let songsList = document.getElementById("album-songsList");
-    for (let song of songs) {
+
+    for (let song of albumsSongs) {
         let songRow = document.createElement("tr");
         songRow.innerHTML = `
                 <td>${song.songTitle}</td>
@@ -105,14 +101,15 @@ albumsSongs.then((songs) => {
         });
         songsList.appendChild(songRow);
     }
-});
-
-deleteAlbumButton.addEventListener("click", () => {
-    deleteAlbum(albumId);
-});
-let allRows = document.querySelectorAll(".table-row");
-allRows.forEach((row) => {
-    row.addEventListener("click", () => {
-        window.location.href = row.getAttribute("data-href");
+    if (deleteAlbumButton) {
+        deleteAlbumButton.addEventListener("click", () => {
+            deleteAlbum(albumId);
+        });
+    }
+    let allRows = document.querySelectorAll(".table-row");
+    allRows.forEach((row) => {
+        row.addEventListener("click", () => {
+            window.location.href = row.getAttribute("data-href");
+        });
     });
-});
+}
